requirements.txt
asyncio
python-qpid-proton
docker